requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.3
plotly==5.18.0
streamlit==1.29.0
openpyxl==3.1.2
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
openai>=1.109.1,<3.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

